Update README.md
A commercial use of the Internet was introduced in the 1980's, but the world did not know that this would change the course of human history. The Internet has opened a new world for many people around the world and it has become a vital part of human life. All the tiring offline jobs are upgraded to online.Since the origin of the internet to today's 5G internet, there have been significant changes in its usage, pricing, and end user base.